Splitting responsibility for AbstractWsgMetricsCollector class
