+++ src/main/java/picard/analysis/AbstractWgsMetricsCollector.java
@@ -89,6 +89,8 @@ public abstract class AbstractWgsMetricsCollector<T extends AbstractRecordAndOff
      */
     protected long counter = 0;
 
+    protected WgsHistogramCollector histogram;
+
     /**
      * Creates a collector and initializes the inner data structures
      *
@@ -106,6 +108,7 @@ public abstract class AbstractWgsMetricsCollector<T extends AbstractRecordAndOff
         this.coverageCap    = coverageCap;
         this.intervals      = intervals;
         this.usingStopAfter = collectWgsMetrics.STOP_AFTER > 0;
+        this.histogram = new WgsHistogramCollector(coverageCap);
     }
 
     /**
@@ -134,33 +137,33 @@ public abstract class AbstractWgsMetricsCollector<T extends AbstractRecordAndOff
 
         // add them to the file
         file.addMetric(metrics);
-        file.addHistogram(getHighQualityDepthHistogram());
-        if (includeBQHistogram) addBaseQHistogram(file);
+        file.addHistogram(histogram.getHighQualityDepthHistogram());
+        if (includeBQHistogram) histogram.addBaseQHistogram(file);
     }
 
-    protected void addBaseQHistogram(final MetricsFile<WgsMetrics, Integer> file) {
-        file.addHistogram(getUnfilteredBaseQHistogram());
-    }
-
-    protected Histogram<Integer> getHighQualityDepthHistogram() {
-        return getHistogram(highQualityDepthHistogramArray, "coverage", "high_quality_coverage_count");
-    }
-
-    protected Histogram<Integer> getUnfilteredDepthHistogram() {
-        return getHistogram(unfilteredDepthHistogramArray, "coverage", "unfiltered_coverage_count");
-    }
-
-    protected Histogram<Integer> getUnfilteredBaseQHistogram() {
-        return getHistogram(unfilteredBaseQHistogramArray, "baseq", "unfiltered_baseq_count");
-    }
-
-    protected Histogram<Integer> getHistogram(final long[] array, final String binLabel, final String valueLabel) {
-        final Histogram<Integer> histogram = new Histogram<>(binLabel, valueLabel);
-        for (int i = 0; i < array.length; ++i) {
-            histogram.increment(i, array[i]);
-        }
-        return histogram;
-    }
+//    protected void addBaseQHistogram(final MetricsFile<WgsMetrics, Integer> file) {
+//        file.addHistogram(getUnfilteredBaseQHistogram());
+//    }
+//
+//    protected Histogram<Integer> getHighQualityDepthHistogram() {
+//        return getHistogram(highQualityDepthHistogramArray, "coverage", "high_quality_coverage_count");
+//    }
+//
+//    protected Histogram<Integer> getUnfilteredDepthHistogram() {
+//        return getHistogram(unfilteredDepthHistogramArray, "coverage", "unfiltered_coverage_count");
+//    }
+//
+//    protected Histogram<Integer> getUnfilteredBaseQHistogram() {
+//        return getHistogram(unfilteredBaseQHistogramArray, "baseq", "unfiltered_baseq_count");
+//    }
+//
+//    protected Histogram<Integer> getHistogram(final long[] array, final String binLabel, final String valueLabel) {
+//        final Histogram<Integer> histogram = new Histogram<>(binLabel, valueLabel);
+//        for (int i = 0; i < array.length; ++i) {
+//            histogram.increment(i, array[i]);
+//        }
+//        return histogram;
+//    }
 
     /**
      * Creates CollectWgsMetrics.WgsMetrics - the object holding the result of CollectWgsMetrics
@@ -176,8 +179,8 @@ public abstract class AbstractWgsMetricsCollector<T extends AbstractRecordAndOff
             final CountingPairedFilter pairFilter) {
         return collectWgsMetrics.generateWgsMetrics(
                 this.intervals,
-                getHighQualityDepthHistogram(),
-                getUnfilteredDepthHistogram(),
+                histogram.getHighQualityDepthHistogram(),
+                histogram.getUnfilteredDepthHistogram(),
                 collectWgsMetrics.getBasesExcludedBy(adapterFilter),
                 collectWgsMetrics.getBasesExcludedBy(mapqFilter),
                 collectWgsMetrics.getBasesExcludedBy(dupeFilter),
@@ -186,7 +189,7 @@ public abstract class AbstractWgsMetricsCollector<T extends AbstractRecordAndOff
                 basesExcludedByOverlap,
                 basesExcludedByCapping,
                 coverageCap,
-                getUnfilteredBaseQHistogram(),
+                histogram.getUnfilteredBaseQHistogram(),
                 collectWgsMetrics.SAMPLE_SIZE);
     }
 
@@ -218,4 +221,5 @@ public abstract class AbstractWgsMetricsCollector<T extends AbstractRecordAndOff
         final byte base = ref.getBases()[position - 1];
         return SequenceUtil.isNoCall(base);
     }
+
 }
